requirements.txt
requests
pyyaml
python-telegram-bot[job-queue]
playwright